move pushSortReadOnlyArgs and SortReadOnlyOptions to generic-transformers
