+++ packages/client/lib/commands/SORT_RO.ts
@@ -1,17 +1,8 @@
+import { pushSortReadOnlyArgs, SortReadOnlyOptions } from "./generic-transformers";
+
 export const FIRST_KEY_INDEX = 1;
 
 export const IS_READ_ONLY = true;
-
-export interface SortReadOnlyOptions {
-    BY?: string;
-    LIMIT?: {
-        offset: number;
-        count: number;
-    },
-    GET?: string | Array<string>;
-    DIRECTION?: 'ASC' | 'DESC';
-    ALPHA?: true;
-}
 
 export function transformArguments(key: string, options?: SortReadOnlyOptions): Array<string> {
     const args = ['SORT_RO', key];
@@ -19,34 +10,6 @@ export function transformArguments(key: string, options?: SortReadOnlyOptions): 
     pushSortReadOnlyArgs(args, options);
 
     return args;
-}
-
-export function pushSortReadOnlyArgs(args: Array<string>, options?: SortReadOnlyOptions) {
-    if (options?.BY) {
-        args.push('BY', options.BY);
-    }
-
-    if (options?.LIMIT) {
-        args.push(
-            'LIMIT',
-            options.LIMIT.offset.toString(),
-            options.LIMIT.count.toString()
-        );
-    }
-
-    if (options?.GET) {
-        for (const pattern of (typeof options.GET === 'string' ? [options.GET] : options.GET)) {
-            args.push('GET', pattern);
-        }
-    }
-
-    if (options?.DIRECTION) {
-        args.push(options.DIRECTION);
-    }
-
-    if (options?.ALPHA) {
-        args.push('ALPHA');
-    }
 }
 
 export declare function transformReply(): Array<string>;
